requirements.txt
# API
fastapi>=0.103.0
uvicorn>=0.23.2
psycopg2-binary>=2.9.7
pydantic>=2.3.0

# Machine Learning
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.25.2
joblib>=1.3.2

# Interface utilisateur
streamlit>=1.26.0
matplotlib>=3.7.2
seaborn>=0.12.2
requests>=2.31.0
plotly>=5.16.1

# Utilitaires
python-dotenv>=1.0.0 

# Tests
pytest>=8.3.5